# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mentora AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type GenerateQuizBodyDifficulty =
  (typeof GenerateQuizBodyDifficulty)[keyof typeof GenerateQuizBodyDifficulty];

export const GenerateQuizBodyDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface GenerateQuizBody {
  topic: string;
  subject: string;
  difficulty: GenerateQuizBodyDifficulty;
  count?: number;
}

export interface QuizQuestion {
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
}

export interface GeneratedQuiz {
  topic: string;
  questions: QuizQuestion[];
}

export type StudentProfileDifficulty =
  (typeof StudentProfileDifficulty)[keyof typeof StudentProfileDifficulty];

export const StudentProfileDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export type StudentProfileLearningStyle =
  (typeof StudentProfileLearningStyle)[keyof typeof StudentProfileLearningStyle];

export const StudentProfileLearningStyle = {
  visual: "visual",
  detailed: "detailed",
  "example-based": "example-based",
  concise: "concise",
} as const;

export interface StudentProfile {
  id: number;
  name: string;
  subjects: string[];
  difficulty: StudentProfileDifficulty;
  learningStyle: StudentProfileLearningStyle;
  createdAt: string;
}

export type CreateProfileBodyDifficulty =
  (typeof CreateProfileBodyDifficulty)[keyof typeof CreateProfileBodyDifficulty];

export const CreateProfileBodyDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export type CreateProfileBodyLearningStyle =
  (typeof CreateProfileBodyLearningStyle)[keyof typeof CreateProfileBodyLearningStyle];

export const CreateProfileBodyLearningStyle = {
  visual: "visual",
  detailed: "detailed",
  "example-based": "example-based",
  concise: "concise",
} as const;

export interface CreateProfileBody {
  name: string;
  subjects: string[];
  difficulty: CreateProfileBodyDifficulty;
  learningStyle: CreateProfileBodyLearningStyle;
}

export interface ProgressSummary {
  totalConversations: number;
  totalMessages: number;
  dayStreak: number;
  averageScore: number;
  topicsStudied: string[];
  quizzesTaken: number;
  hoursLearned: number;
}

export interface SaveQuizResultBody {
  topic: string;
  subject: string;
  score: number;
  totalQuestions: number;
}

export interface QuizResult {
  id: number;
  topic: string;
  subject: string;
  score: number;
  totalQuestions: number;
  createdAt: string;
}