# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mentora AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get student profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    post:
      operationId: createProfile
      tags: [profile]
      summary: Create student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
...
components:
  schemas:
...
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subjects:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        learningStyle:
          type: string
          enum: [visual, detailed, example-based, concise]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - subjects
        - difficulty
        - learningStyle
        - createdAt
    CreateProfileBody:
      type: object
      properties:
        name:
          type: string
        subjects:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        learningStyle:
          type: string
          enum: [visual, detailed, example-based, concise]
      required:
        - name
        - subjects
        - difficulty
        - learningStyle
...