# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mentora AI API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Generate a quiz for a given topic
 */
export const GenerateQuizBody = zod.object({
  topic: zod.string(),
  subject: zod.string(),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  count: zod.number().optional(),
});

export const GenerateQuizResponse = zod.object({
  topic: zod.string(),
  questions: zod.array(
    zod.object({
      question: zod.string(),
      options: zod.array(zod.string()),
      correctIndex: zod.number(),
      explanation: zod.string(),
    }),
  ),
});

/**
 * @summary Get student profile
 */
export const GetProfileResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  subjects: zod.array(zod.string()),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  learningStyle: zod.enum(["visual", "detailed", "example-based", "concise"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Create student profile
 */
export const CreateProfileBody = zod.object({
  name: zod.string(),
  subjects: zod.array(zod.string()),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  learningStyle: zod.enum(["visual", "detailed", "example-based", "concise"]),
});

/**
 * @summary Update student profile
 */
export const UpdateProfileBody = zod.object({
  name: zod.string(),
  subjects: zod.array(zod.string()),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  learningStyle: zod.enum(["visual", "detailed", "example-based", "concise"]),
});

export const UpdateProfileResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  subjects: zod.array(zod.string()),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  learningStyle: zod.enum(["visual", "detailed", "example-based", "concise"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get learning progress summary
 */
export const GetProgressResponse = zod.object({
  totalConversations: zod.number(),
  totalMessages: zod.number(),
  dayStreak: zod.number(),
  averageScore: zod.number(),
  topicsStudied: zod.array(zod.string()),
  quizzesTaken: zod.number(),
  hoursLearned: zod.number(),
});

/**
 * @summary Save a quiz result
 */
export const SaveQuizResultBody = zod.object({
  topic: zod.string(),
  subject: zod.string(),
  score: zod.number(),
  totalQuestions: zod.number(),
});

/**
 * @summary List all quiz results
 */
export const ListQuizResultsResponseItem = zod.object({
  id: zod.number(),
  topic: zod.string(),
  subject: zod.string(),
  score: zod.number(),
  totalQuestions: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListQuizResultsResponse = zod.array(ListQuizResultsResponseItem);