# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mentora AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentProfileDifficulty } from "./studentProfileDifficulty";
import type { StudentProfileLearningStyle } from "./studentProfileLearningStyle";

export interface StudentProfile {
  id: number;
  name: string;
  subjects: string[];
  difficulty: StudentProfileDifficulty;
  learningStyle: StudentProfileLearningStyle;
  createdAt: Date;
}